requirements.txt
pymongo
pandas
scipy
matplotlib
python-dotenv